+++ src/dialogs/settingsdialog.h
@@ -243,6 +243,8 @@ class SettingsDialog : public MasterDialog {
 
     void on_markdownLspEnabledCheckBox_toggled(bool checked);
 
+    void on_enableWikiLinkSupportCheckBox_toggled(bool checked);
+
     void onLayoutSettingsStored(const QString &workspaceIdentifier);
 
     void on_localTrashEnabledCheckBox_toggled(bool checked);